Move Linux kernel module support
Separate out the support for kernel modules on the Linux backend into
linux_kernel_module.go
